Stop installing SwiftLint when it's not installed (#246)
The SwiftLint script run phase used backticks around "brew bundle" when
printing a message when SwiftLint is not installed. In shell,
expressions in backticks are evaluated, or in other words, executed.
So every time this script ran on a machine without Swiftlint installed
it basically install SwiftLint. This is not a desired behavior.